Feat: 라인업 컴포넌트 tailwind로 변경
